perform box score calculations

+++ NBA.Service/BoxScoreEvaluation.cs
@@ -4,17 +4,17 @@ namespace NBA.Service
 {
     public static class BoxScoreEvaluation
     {
-        public const double Points = 1.0;
-        public const double Assists = 1.5;
-        public const double Rebounds = 1.2;
-        public const double Blocks = 3.0;
-        public const double Steals = 3.0;
-        public const double ThreePointsMade = 1.0;
-        public const double ThreePointsMissed = -1.0;
-        public const double Turnovers = -2.0;
-        public const double FreeThrowMade = 1.0;
-        public const double FreeThrowMissed = -1.0;
-        public const double FieldGoalMade = 1.0;
-        public const double FieldGoalMissed = -1.0;
+        public const decimal Points = 1.0m;
+        public const decimal Assists = 1.5m;
+        public const decimal Rebounds = 1.2m;
+        public const decimal Blocks = 3.0m;
+        public const decimal Steals = 3.0m;
+        public const decimal ThreePointsMade = 1.0m;
+        public const decimal ThreePointsMissed = -1.0m;
+        public const decimal Turnovers = -2.0m;
+        public const decimal FreeThrowMade = 1.0m;
+        public const decimal FreeThrowMissed = -1.0m;
+        public const decimal FieldGoalMade = 1.0m;
+        public const decimal FieldGoalMissed = -1.0m;
     }
 }
